created another component activityListItem
