+++ client/src/features/activities/dashboard/ActivityList.tsx
@@ -1,45 +1,18 @@
 import { observer } from 'mobx-react-lite';
-import React, { SyntheticEvent, useState } from 'react';
-import { Link } from 'react-router-dom';
-import { Button, Item, Label, Segment } from 'semantic-ui-react';
+import React, {  } from 'react';
+import { Item, Segment } from 'semantic-ui-react';
 import { useStore } from '../../../app/stores/store';
+import ActivityListItem from './ActivityListItem';
 
 export default observer(function ActivityList() {
     const {activityStore} = useStore();
-    const {deleteActivity, activitiesByDate, loading} = activityStore;
-
-    const [target, setTarget] = useState('');
-    
-    function handleActivityDelete(e: SyntheticEvent<HTMLButtonElement>, id: string) {
-        setTarget(e.currentTarget.name);
-        deleteActivity(id);
-    }
+    const {activitiesByDate} = activityStore;
 
     return (
         <Segment>
             <Item.Group divided>
                 {activitiesByDate.map(activity => (
-                    <Item key={activity.id}>
-                        <Item.Content>
-                            <Item.Header as='a'>{activity.title}</Item.Header>
-                            <Item.Meta>{activity.date}</Item.Meta>
-                            <Item.Description>
-                                <div>{activity.description}</div>
-                                <div>{activity.city}, {activity.venue}</div>
-                            </Item.Description>
-                            <Item.Extra>
-                                <Button as={Link} to={`/activities/${activity.id}`} floated='right' content='View' color='blue' />
-                                <Button
-                                name={activity.id} 
-                                loading={loading && target === activity.id} 
-                                onClick={(e) => handleActivityDelete(e, activity.id)}
-                                floated='right' 
-                                content='Delete' 
-                                color='red' />
-                                <Label basic content={activity.category} />
-                            </Item.Extra>
-                        </Item.Content>
-                    </Item>
+                    <ActivityListItem key={activity.id} activity={activity} />
                 ))}
             </Item.Group>
         </Segment>
